Cleaned up documentation.
Removed all mentions of optional color degradation, since the palette
which enables that was removed in the previous commit.

Removed all pleading for setting your terminal colors, since the colors
are now based on the 256-color palette.

Removed the value table since it's already present in the code.

Removed trailing whitespace.